requirements.txt
google-generativeai
numpy
pandas
faiss-cpu
